% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchCalibrate.R
\name{BatchCalibrate}
\alias{BatchCalibrate}
\title{Batch Radio Transmitter Calibratiion.}
\usage{
BatchCalibrate(data)
}
\arguments{
\item{data}{a dataset with 5 columns: "transmitter", "duration", "pulses", "period", and "temperature.}
}
\value{
Will generate a table with both a linear and quadratic equation for the relationship between temperature (Y variable) and period (X variable). A scatterplot illustrating both models will also be generated and saved. The function anova() is used to compare the fit of both models. A statement will be also be generated that details which model is a better fit, and the p-value for the model comparison test will be provided.
}
\description{
Calibrates one or more radio transmitters. Requires a dataset with 5 columns: "transmitter", "duration", "pulses", "period", and "temperature. BatchCalibrate() will generate a file named "batch_calibration.csv" with 1 row for each transmitter. For each transmitter, a linear equation, a quadratic equation, a statement that details which model is a better fit, and the p-value for the model comparison test will be provided. The function anova() is used as the model comparison test. The "batch_calibration.csv" file will be saved to the current working directory. Within the working directory must also exist a folder titled "calibration_images". A jpeg file of a scatterplot (temperature~period) for each transmitter with both linear and quadratic regression lines will be saved within the "calibration_images" folder.
}
