% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempCalibrate.R
\name{TempCalibrate}
\alias{TempCalibrate}
\title{Temperature Calibration of Radio Rransmitters.}
\usage{
TempCalibrate(data)
}
\arguments{
\item{data}{a dataset that contains a column titled "period" and a column titled "temperature".}
}
\value{
Will generate  both a linear and quadratic equation for the relationship between temperature (Y variable) and period (X variable). A scatterplot illustrating both models will also be generated. The function anova() is used to compare the fit of both models. A statement will be also be generated that details which model is a better fit, and the p-value for the model comparison test will be provided.
}
\description{
Requires a dataset that contains a column titled "period" and a column titled "temperature".
}
